TEAM7-90 #Similar authors page

+++ Phase-4/src/main/java/com/team7/FavoriteList.java
@@ -19,6 +19,7 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
+import javax.swing.ListSelectionModel;
 import javax.swing.SwingConstants;
 import javax.swing.border.EmptyBorder;
 import javax.swing.table.DefaultTableCellRenderer;
@@ -30,12 +31,12 @@ public class FavoriteList extends JFrame {
 
 	private static final long serialVersionUID = 1L;
 	private JPanel contentPane;
-	JButton btnNewButton;
+	JButton btnLogout;
 	JButton btnSearch;
 	JTable table;
-	private JButton btnRemove;
+	JButton btnRemove;
 
-	public FavoriteList() throws IOException {
+	public FavoriteList() {
 		
 		setVisible(true);
 		setTitle("Favorite List");
@@ -73,8 +74,8 @@ public class FavoriteList extends JFrame {
 			}
 		});
 
-		btnNewButton = new JButton("LogOut");
-		btnNewButton.addActionListener(new ActionListener() {
+		btnLogout = new JButton("LogOut");
+		btnLogout.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
 				ButtonEditor.savedAuthors.clear();
 				// make the currentUser null and redirect to login page
@@ -87,9 +88,9 @@ public class FavoriteList extends JFrame {
 				log.setLocationRelativeTo(null);
 			}
 		});
-		btnNewButton.setFont(new Font("Lucida Grande", Font.BOLD, 16));
-		btnNewButton.setBounds(1253, 13, 117, 34);
-		panel.add(btnNewButton);
+		btnLogout.setFont(new Font("Lucida Grande", Font.BOLD, 16));
+		btnLogout.setBounds(1253, 13, 117, 34);
+		panel.add(btnLogout);
 
 		JPanel panel_1 = new JPanel();
 		panel_1.setBounds(0, 119, 1382, 330);
@@ -100,6 +101,7 @@ public class FavoriteList extends JFrame {
 			ImplementSchemaDB db =  new ImplementSchemaDB();
 			Connection conn = db.getConnection();
 
+			// statements to go to the query engine
 			PreparedStatement p = conn.prepareStatement("select selectedAuthor from Favorite_list where username='"+LoginUI.currentUser+"'");
 			ResultSet rs = p.executeQuery();
 
@@ -121,36 +123,46 @@ public class FavoriteList extends JFrame {
 
 			table.setPreferredScrollableViewportSize(new Dimension(400, 300));
 
+			table.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
+			
 			JScrollPane scroll = new JScrollPane(table);
 			setVisible(true);
 
 			panel_1.add(scroll);
-			DefaultTableModel mod = (DefaultTableModel) table.getModel();
+			DefaultTableModel model = (DefaultTableModel) table.getModel();
 
 			btnRemove = new JButton("Remove");
 			btnRemove.addActionListener(new ActionListener() {
 				public void actionPerformed(ActionEvent arg0) {
 
 					int index = table.getSelectedRow();
-					String author = (String) mod.getValueAt(index, 0);
-					ImplementSchemaDB db =  new ImplementSchemaDB();
-					Connection conn;
-					try {
-						conn = db.getConnection();
-						PreparedStatement stmt = conn.prepareStatement("Delete from Favorite_list where selectedAuthor=?");								
-						stmt.setString(1,author);							
-						stmt.executeUpdate();		
-						mod.removeRow(index);; 
-					} catch (IOException | SQLException e) {
-						e.printStackTrace();
+					if (index == -1) {
+						LoginUI log = new LoginUI();
+						log.messageShow("Please select an author to remove");
 					}
+					else{
+						String author = (String) model.getValueAt(index, 0);
+						ImplementSchemaDB db =  new ImplementSchemaDB();
+						Connection conn;
+						try {
+							conn = db.getConnection();
+							PreparedStatement stmt = conn.prepareStatement("Delete from Favorite_list where selectedAuthor=?");								
+							stmt.setString(1,author);							
+							stmt.executeUpdate();		
+							model.removeRow(index);; 
+						} catch (IOException | SQLException e) {
+							e.printStackTrace();
+						}
+					}
+
 
 
 				}
 			});
 			btnRemove.setBounds(643, 486, 97, 30);
 			contentPane.add(btnRemove);
-		} catch (SQLException e2) {
+		} catch (SQLException | IOException e2) {
+			
 		}
 	}
